D2 - find_person, find_entertainment
Both are partially done. First part is done (when only a part of a name or movie/series is written in the prompt). Second part is missing (a possible misspelling should still return either people or movies/title. In this scenario, the order of which possible matches should occur should not be random. Rather, the most popular movie/series, popular actress ect. should occur first).