SearchResults
Name of title is now shown instead of its id. A block of code is generalized, so that both name of title and name of series is displayed. For that reason, SearchController and SearchResultModel has also been modified.
If we modify our search algorithm, so that a user can also search for a person, a similar block of code has also been added to SearchReaults.js.